Extracted base implementations
